requirements.txt
python-dotenv==1.0.0
openai>=1.66.0
openai-agents
streamlit>=1.29.0
pydantic>=2.0.0